+++ govtool/frontend/src/components/organisms/TopNav.tsx
@@ -8,7 +8,7 @@ import { useCardano, useFeatureFlag, useModal } from "@context";
 import { useScreenDimension, useTranslation } from "@hooks";
 import { openInNewTab } from "@utils";
 import { DrawerMobile } from "./DrawerMobile";
-import { MaintenanceEndingBanner } from "./MaintenanceEndingBanner";
+import { useMaintenanceEndingBannerContext } from "./MaintenanceEndingBanner";
 
 const POSITION_TO_BLUR = 80;
 
@@ -25,6 +25,9 @@ export const TopNav = ({ isConnectButton = true }) => {
   const { isEnabled, disconnectWallet, stakeKey } = useCardano();
   const navigate = useNavigate();
   const { t } = useTranslation();
+
+  const { height: maintenanceEndingBannerHeight } =
+    useMaintenanceEndingBannerContext();
 
   useEffect(() => {
     const onScroll = () => {
@@ -137,9 +140,14 @@ export const TopNav = ({ isConnectButton = true }) => {
   );
 
   return (
-    <Box sx={{ position: "sticky", top: 0, zIndex: 100, width: "100%" }}>
-      <MaintenanceEndingBanner />
-
+    <Box
+      sx={{
+        position: "sticky",
+        top: maintenanceEndingBannerHeight,
+        zIndex: 100,
+        width: "100%",
+      }}
+    >
       <AppBar
         ref={containerRef}
         position="static"