+++ govtool/frontend/src/types/@intersect.mbo.d.ts
@@ -5,7 +5,7 @@ enum MetadataValidationStatus {
   INCORRECT_FORMAT = "INCORRECT_FORMAT",
 }
 declare module "@intersect.mbo/pdf-ui/cjs" {
-  import { EpochParams } from "@/models";
+  import { EpochParams, Account } from "@/models";
 
   type ProposalDiscussionProps = {
     pdfApiUrl: string;
@@ -32,6 +32,16 @@ declare module "@intersect.mbo/pdf-ui/cjs" {
     epochParams?: EpochParams;
     username: string;
     setUsername: (username: string) => void;
+    getAdaHolderVotingPower: ({
+      stakeKey,
+    }: {
+      stakeKey?: string;
+    }) => Promise<number>;
+    getAccount: ({
+      stakeKey,
+    }: {
+      stakeKey?: string;
+    }) => Promise<Account>;
   };
 
   type GovernanceActionsOutcomesProps = {