+++ tests/govtool-frontend/playwright/lib/helpers/cardano.ts
@@ -50,7 +50,7 @@ export async function skipIfNotInfoAndBootstrapping(type: ProposalType) {
 export async function skipIfMainnet() {
   if (environments.networkId === 1) {
     await allure.description(
-      "Ada spendable features are not available on mainnet."
+      "Test skipped on mainnet to prevent affecting the production environment or using real ADA."
     );
     test.skip();
   }